Fixes parabolas wrong return values
The previous implementation of Simpson.parabolas didn't return the
correct values because it merely iterated through each point. Each point
is indeed iterated through, however some are iterated through more than
once. The fix was to iterate through the partitions of the interval and
get the points needed from the partitions.