Very basic crappy animation implementation
Each graph method of Quadrature now returns a list of artists which then
get passed to the animation method of Graph, which then gets passed to
FuncAnimation which then gets saved as a file. I'm sure there's a better
way to do this, but that's a problem for future me.